Feat: fix lsp add cheatcheat to tmux

+++ .build/cppbuild/buil.sh
@@ -10,7 +10,6 @@ then
 	echo "build srcs"
 	echo "build includes"
 	echo "build build"
-	echo "build Makefile"
 	echo "build .ccls"
 	echo "build CMakeLists"
 	echo "build main.cpp"
@@ -19,7 +18,12 @@ then
 
 	rm -rf build/*
 	mkdir build srcs includes
-	cp ~/.build/cppbuild/Makefile .
+	if [ -f "Makefile" ]; then
+		echo "Makefile exits"
+	else
+		echo "build Makefile"
+		cp ~/.build/cppbuild/Makefile .
+	fi
 	cp ~/.build/cppbuild/.ccls .
 	touch srcs/main.cpp 
 	touch srcs/$2.cpp 
@@ -33,14 +37,18 @@ then
 	echo "build srcs"
 	echo "build includes"
 	echo "build build"
-	echo "build Makefile"
 	echo "build .ccls"
 	echo "build CMakeLists"
 	echo "build main.cpp"
 
 	rm -rf build/*
 	mkdir build srcs includes
-	cp ~/.build/cppbuild/Makefile .
+	if [ -f "Makefile" ]; then
+		echo "Makefile exits"
+	else
+		echo "build Makefile"
+		cp ~/.build/cppbuild/Makefile .
+	fi
 	cp ~/.build/cppbuild/.ccls .
 	sed "s/name_project/$1/g" ~/.build/cppbuild/CMakeLists.txt > CMakeLists.txt
 	sed "s/bin/$1/g" ~/.build/cppbuild/.vimspector.json > .vimspector.json
